refactor: 공연 관리자 서비스 클래스 명 통일
